Bring some sanity into the world
Use async functions to do async things with executing commands
to detect languages and such. Also detect all languages before
doing anything.

Switching 'execution views' based on the language is still ugly
as sin though, and needs to be re-thought.

Also maybe having execution views at all is premature, we can just
stick with having a terminal for all of them and only customize
the options available in the toolbar?