Preventing terminal view flickering
Preventing terminal view flickering by changing how the state data
flows. Instead of performing state changes after the ProgressView is
presented, we now perform the necessary state changes when other state
changes.

To achieve this in a easier way, I had to introduce a new state
`languageStatus` which only represents the status of the language
terminal.
I also removed languages state, because the languages it self does not need
to change and can stay static.